Hibernate implemented

+++ src/main/java/io/github/mat3e/HelloService.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 class HelloService {
     public static final String FALLBACK_NAME = "world";
-    public static final Lang FALLBACK_LANG = new Lang(1L, "Hello", "en");
+    public static final Lang FALLBACK_LANG = new Lang(1, "Hello", "en");
     private final Logger logger = LoggerFactory.getLogger(HelloService.class);
     private LangRepository repository;
 
@@ -20,9 +20,9 @@ class HelloService {
     }
 
     String prepareGreeting(String name, String lang) {
-        Long langId;
+        Integer langId;
         try {
-            langId= Optional.ofNullable(lang).map(Long::valueOf).orElse(FALLBACK_LANG.getId());
+            langId= Optional.ofNullable(lang).map(Integer::valueOf).orElse(FALLBACK_LANG.getId());
         } catch (NumberFormatException e) {
             logger.warn("Non-numeric language id used: " + lang);
             langId = FALLBACK_LANG.getId();
